Day 11
Well, plain brute force fails for the first time this year...
So adding little 'saving' mechanic using array that stores already
counted values of "how many values an X value creates in Y blinks"
that can be reused.

Also adding -lm to testall.sh as it's required for this code.
I guess it's not harmful to add it to all days even if they dont use
math.h at all.