Day 14
Part 1 is calculated on the fly with modulo operation. For Part 2,
I am calculating and storing all new positions of points and then
calculating mean distance between them. If this value is much lower
(less than 60% - I had to get this number from logs), we have encountered
some anomaly. If program is given anything as 2nd parameter, it will
also draw the image.