Day 19
In part 1 I used trick to first check if each towel can be made of
other towels to reduce amount of searching possibilities... only to
find out in part 2 that we are interested in ALL of the possibilities. :D

So then I removed that in part 2 code but instead added hash mapping
from <search.h> to cache already calculated towels.

Also in both parts there is my own function to check if one string
starts with another that after finishing successfully returns pointer
past the checked string.