Day 23
Quite satisfied with today's solution. Already in Part 1 I decided
to store input information in 2 different data structures at once:
1) char link[26][26][26][26] that stores information if connection
   between 2 PCs exists

2) List of PCs that store their names and list of PCs they are
   connected with.

With those I was able in Part 1 just iterate over every PC and check
every pair from its connection list to see if it's connected itself.

And in Part 2 I extended it into checking not pair but every connection
and add it to list if it is connected to every already present PC from list.